% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{otp_check_mode}
\alias{otp_check_mode}
\title{Check and process transport mode}
\usage{
otp_check_mode(mode)
}
\arguments{
\item{mode, }{character vector}
}
\description{
Check and process transport mode
}
\keyword{internal}
