% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{otp_check_params}
\alias{otp_check_params}
\title{Check OTP parameters}
\usage{
otp_check_params(otpcon, ...)
}
\arguments{
\item{otpcon}{Object of otpcon class}

\item{...}{Other optional parameters}
}
\description{
Check OTP parameters
}
\keyword{internal}
