% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otpr_create_surface.R
\name{otp_create_surface}
\alias{otp_create_surface}
\title{Creates a travel time surface (OTPv1 only).}
\usage{
otp_create_surface(
  otpcon,
  getRaster = FALSE,
  rasterPath = tempdir(),
  origin,
  mode = "TRANSIT",
  date,
  time,
  maxWalkDistance = 800,
  walkReluctance = 2,
  transferPenalty = 0,
  minTransferTime = 0
)
}
\arguments{
\item{otpcon}{An OTP connection object produced by \code{\link{otp_connect}}.}

\item{getRaster}{Logical. Whether or not to download a raster (geoTIFF) of the generated
surface. Default FALSE.}

\item{rasterPath}{Character. Path of a directory where the the surface raster
should be saved. Default is \code{tempdir()}. The file will be named 
surface_{id}.tiff, with {id} replaced by the OTP id assigned to the surface.}

\item{origin}{Numeric vector, Latitude/Longitude pair, e.g. `c(53.48805, -2.24258)`}

\item{mode}{Character, mode of travel. Valid values are: WALK, TRANSIT, BUS,
or RAIL.
Note that WALK mode is automatically included for TRANSIT, BUS and RAIL.
TRANSIT will use all available transit modes. Default is TRANSIT.}

\item{date}{Character, must be in the format mm-dd-yyyy. This is the desired
date of travel. Only relevant if \code{mode} includes public transport.
Default is current system date.}

\item{time}{Character, must be in the format hh:mm:ss. This is the desired 
departure time. Default is current system time.}

\item{maxWalkDistance}{Numeric. The maximum distance (in meters) the user is
willing to walk. Default = 800.}

\item{walkReluctance}{Integer. A multiplier for how bad walking is, compared
to being in transit for equal lengths of time. Default = 2.}

\item{transferPenalty}{Integer. An additional penalty added to boardings after
the first. The value is in OTP's internal weight units, which are roughly equivalent
to seconds. Set this to a high value to discourage transfers. Default is 0.}

\item{minTransferTime}{Integer. The minimum time, in seconds, between successive
trips on different vehicles. This is designed to allow for imperfect schedule
adherence. This is a minimum; transfers over longer distances might use a longer time.
Default is 0.}
}
\value{
A list of 5 items:
\itemize{
\item \code{errorId} - character, should be "OK" if no error condition.
\item \code{query} - this is a character string containing the URL.
that was submitted to the OTP API.
\item \code{surfaceId} - integer, the id of the surface that was evaluated.
\item \code{surfaceRecord} - details of the parameters used to create the surface.
\item \code{rasterDownload} - the path to the saved raster file (if \code{getRaster} was 
set to TRUE and a valid path was provided via \code{rasterPath}.)
}
}
\description{
Creates a travel time surface for an origin point. A surface contains the travel time
to every geographic coordinate that can be reached from that origin. Optionally, the surface
can be saved as a raster file (GeoTIFF) to the designated directory.
}
\examples{
\dontrun{
otp_create_surface(otpcon, origin = c(53.43329,-2.13357), mode = "TRANSIT", 
maxWalkDistance = 1600, getRaster = TRUE)

otp_create_surface(otpcon, origin = c(53.43329,-2.13357), date = "03-26-2019",
time = "08:00:00", mode = "BUS", maxWalkDistance = 1600, getRaster = TRUE)
}
}
