% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp_evaluate_surface.R
\name{otp_evaluate_surface}
\alias{otp_evaluate_surface}
\title{Evaluates an existing surface. Using a pointset from the specified CSV file,
the travel time to each point is obtained from the specified surface. Accessibility
indicators are then generated for one or more 'opportunity' columns in the pointset
file. For example, you may have the number of jobs available at each location, or
the number of hospital beds.}
\usage{
otp_evaluate_surface(otpcon, surfaceId, pointset, detail = FALSE)
}
\arguments{
\item{otpcon}{An OTP connection object produced by \code{\link{otp_connect}}.}

\item{surfaceId}{Integer, The id number of an existing surface for an origin
created using \code{otp_create_surface()}.}

\item{pointset}{Character string, the name of a pointset known to OTP. A pointset
is contained in a CSV file located in the pointset directory passed to OTP at startup.
The name of the pointset is the name of the file.}

\item{detail}{logical, whether the travel time to each location in the pointset
should be returned. Default is FALSE.}
}
\description{
Requires OTP to have been started with the --analyst switch and the --pointset
parameter set to the path of a directory containing the pointset file(s).
}
\examples{
\dontrun{

}
}
