% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{otp_from_epoch}
\alias{otp_from_epoch}
\title{Convert time from EPOCH format}
\usage{
otp_from_epoch(epoch, tz)
}
\arguments{
\item{epoch, }{time since EPOCH (milliseconds)}

\item{tz, }{timezone (string)}
}
\description{
Convert time from EPOCH format
}
