% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{otp_is_date}
\alias{otp_is_date}
\title{Checks the date format}
\usage{
otp_is_date(date)
}
\arguments{
\item{date}{the supplied date.}
}
\description{
Checks the date format
}
\keyword{internal}
