% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{otp_vector_match}
\alias{otp_vector_match}
\title{Checks if two vectors are the same but where order doesn't matter}
\usage{
otp_vector_match(a, b)
}
\arguments{
\item{a}{vector, to be matched}

\item{b}{vector, to be matched}
}
\description{
Checks if two vectors are the same but where order doesn't matter
}
\keyword{internal}
