% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{otpr_vectorMatch}
\alias{otpr_vectorMatch}
\title{Checks if two vectors are the same but where order doesn't matter}
\usage{
otpr_vectorMatch(a, b)
}
\arguments{
\item{a}{vector, to be matched}

\item{b}{vector, to be matched}
}
\description{
Checks if two vectors are the same but where order doesn't matter
}
