% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otpr_create_surface.R
\name{otp_create_surface}
\alias{otp_create_surface}
\title{Creates a travel time surface for an origin point. A surface has the travel time
to every geographic coordinate that can be reached from that origin. (OTPv1 only)}
\usage{
otp_create_surface(
  otpcon,
  getRaster = FALSE,
  rasterPath = tempdir(),
  location,
  mode = "TRANSIT",
  date,
  time,
  arriveBy = FALSE,
  maxWalkDistance = 800,
  walkReluctance = 2,
  transferPenalty = 0,
  minTransferTime = 0
)
}
\arguments{
\item{otpcon}{An OTP connection object produced by \code{\link{otp_connect}}.}

\item{getRaster}{Logical. Whether to download a raster (tiff) of the generated
surface. Default FALSE.}

\item{rasterPath}{Character. Path of a directory to save the surface raster.
Default is \code{tempdir()}. File will be named surface_{id}.tiff, with {id}
replaced with the id of the surface.}

\item{location}{Numeric vector, Latitude/Longitude pair, e.g. `c(53.48805, -2.24258)`}

\item{mode}{Character, mode of travel. Valid values are: WALK, TRANSIT, BUS,
or RAIL.
Note that WALK mode is automatically included for TRANSIT, BUS and RAIL.
TRANSIT will use all available transit modes. Default is TRANSIT.}

\item{date}{Character, must be in the format mm-dd-yyyy. This is the desired
date of travel. Only relevant if \code{mode} includes public transport.
Default is current system date.}

\item{time}{Character, must be in the format hh:mm:ss. If \code{arriveBy} is
FALSE (the default) this is the desired departure time, otherwise the desired
arrival time. Default is current system time.}

\item{arriveBy}{Logical. Whether the specified date and time is for
departure (FALSE) or arrival (TRUE). Default is FALSE.}

\item{maxWalkDistance}{Numeric. The maximum distance (in meters) the user is
willing to walk. Default = 800.}

\item{walkReluctance}{Integer. A multiplier for how bad walking is, compared
to being in transit for equal lengths of time. Default = 2.}

\item{transferPenalty}{Integer. An additional penalty added to boardings after
the first. The value is in OTP's internal weight units, which are roughly equivalent
to seconds. Set this to a high value to discourage transfers. Default is 0.}

\item{minTransferTime}{Integer. The minimum time, in seconds, between successive
trips on different vehicles. This is designed to allow for imperfect schedule
adherence. This is a minimum; transfers over longer distances might use a longer time.
Default is 0.}
}
\value{
Returns ...
}
\description{
Returns information about the created surface.
}
\examples{
\dontrun{

}
}
